openafs-rpm-package-build: remove files between downloads
Ansible's "uri:" download looks at the file and sets the
last-modified-time header, basically assuming if the file with the
download name on disk is later than the upstream version it doesn't
need to download.  This isn't the case for these downloads, where we
want each one as they are separate clients.

Remove the downloaded file between extractions.

Change-Id: I671e9df072a6c59d29659a65829af18deb5e3c9a