build-wheel-cache: Reduce arm64 builds and various small updates
Because there are many more wheels to build on arm64, we need to
reduce the arm64 builds to just latest two branches to have a
reasonable build time.

Output the script output to logs; if the job gets timeout killed we
still get some log output.

Remove --progress from parallel as it just floods the logs with output
that isn't useful to store.

Remove --verbose from pip to try and reduce logs.  We can still see
failures but currently they are like 500mb compressed.

Zip and remove the logs in post so we don't copy over uncompressed
logs if we are timeout killed.

Change-Id: I5fba52af7fcafb12d0d2c4b416d5e0b7e4feb238