Remove greande jobs for EM and oldest stable
As discussed in ML thread[1], we are going to
make grenade jobs as non voting for all EM stable and
oldest stable. grenade jobs are failing not and it might take
time to fix those if we are able to fix. Once it jobs are
working depends on project team, they can bring them back to
voting or keep non-voting.

With n-v jobs on projects side we need to udpate the integrated
gate template to remove the job from gate pipeline.

Needed-By: https://review.opendev.org/#/c/737414/1
Needed-BY: https://review.opendev.org/#/c/737332/
Needed-By: https://review.opendev.org/#/c/737408/
[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015499.html

Change-Id: I0a924cb298bd74034a9b33d31d2fd506391ac16d