Add template for charm check and gate
Add new template to cover the required range of py3 versions
that the OpenStack Charms need to be validated against.

This is mainly driven by the Ubuntu versions the charms have
to support which includes back to Ubuntu 16.04 LTS (Python 3.5)
and to the current 20.04 LTS (Python 3.8).

Master branch of all charms will be switched to use this
template in time for the next charm release in August 2020.

Stable branches will not be updated however old templates
have been marked as DEPRECATED FOR REMOVAL.

Change-Id: Ib9ed240aef6680be7c972088ca5337b5740b9c36