Revert "Run openstack-tox-lower-constraints on bionic"
This reverts commit 9a491100278eccc40219f51a329c0d0980a3d8a6.

As per the Focal migration work, most of the projects have been fixed
for running the l-c job on Focal[1].

As part of this migration, l-c testing the min version of deps can work
on tested Distro which is Focal in this. If I am installing Nova on Focal
then I will be ineterested to see what all min version of Nova deps can
work in Focal. We can debate on this if we need to test on any older distro
which as per supported python or any other distro and accordingly we can
test the l-c job accordingly. But for this in middle of migration we should
keep it tested on Focal compatible.

Cinder and Keystone projects are two main project failing on l-c, let's
keep running their l-c jobs in Bionic
- cinder: https://review.opendev.org/#/c/751118/
- keystone: https://review.opendev.org/#/c/751122/


[1] http://lists.openstack.org/pipermail/openstack-discuss/2020-September/017155.html
Change-Id: I34ef635d1987bde03fe5944f511dfe7ec1fb741f