openstack-tox-docs: Enable PDF build for stein-ussuri xenial job
https://review.opendev.org/#/c/738326/ created openstack-tox-docs
job for stein-ussuri branches, but it did not specify pre-run and
run phases required for PDF build.
Repositories which enable PDF build with RSVG support requires
librsvg2-bin even for HTNL build (as rsvg-converter is always
specified in doc/source/conf.py). As a result, the above change
breaks HTML doc build in stein-ussuri branches for such repos.
An example is found at [1] (networking-ovn stable/train branch).
This commit makes openstack-tox-docs stein-ussuri same as
that for the master branch.

[1] https://zuul.opendev.org/t/openstack/build/0cd86df4b6aa4896ba033d68d4357948

Change-Id: I61ed9d13f39e88a433d0db2feb5225a1bec861d2