translation: Handle renaming of Chinese locales in Django
Horizon is switching Chinese locales from zh-cn/zh-tw to zh-hans/zh-hant
in horizon and horizon plugins.
This change follows Django change which happend more than five years ago
and the new language code form is commonly used.

Language codes in Zanata are not changed, so we need to covert the
relevant language codes in propose_translation_update.sh.
It allows us to avoid doing it manually and also allows us to
switch these language codes almost once across horizon and plugins.

The detail is explained in the mailing list post.
http://lists.openstack.org/pipermail/openstack-discuss/2021-February/020169.html

Change-Id: I1420508f0513c52866aa637784a51f7c22f6d204