Remove neutron and networking-midonet Ocata jobs definitions
This patch removes legacy-neutron-dsvm-api job definition
It seems that this job was used only in the Neutron stable/Ocata branch
which is going to be EOL with [1].
In the stable/Pike branch Neutron is using only
"legacy-neutron-dsvm-api-pecan" job which isn't removed by that patch.

It also removes networking-midonet jobs:
* periodic-tempest-dsvm-networking-midonet-aio-ml2-full-ocata
* periodic-tempest-dsvm-networking-midonet-aio-v2-full-ocata

As those jobs were also used only in Ocata branch which is EOL in
networking-midonet already.

[1] https://review.opendev.org/c/openstack/releases/+/790904

Change-Id: I7b9e6cb20d7b8d8d1dff9b8bee2c652df743b2ba