Define publish-to-pypi-stable-only template
Deprecated deliverables only do the release on
stable branches only not on master. But existing
publish-to-pypi template does not have the branch
matcher to exclude the master gate from running the
release jobs.

This commit defines a new template for such deprecated
repo publish-to-pypi-stable-only which exclude the master
branch to run the release jobs.

This new template will be used in all the deprected repo so
that if needed we can update the master conent which is README.rst
for example: https://review.opendev.org/c/openstack/neutron-fwaas/+/800140

Change-Id: Icc98e4f231413365af6a4f2270282565335e343c