Remove retired neutron-lbaas and django-openstack-auth from job
neutron-lbaas is retired now and we can
remove it from zuul jobs. It also retire django_openstack_auth
in this patch to unblock the gate.

`django_openstack_auth`` code was merged into the main horizon
repository during the queens release. This repo is already
deprecated and now we can retire it. This patch removes all
references of ``django_openstack_auth`` as mentioned in [1].

[1] https://docs.openstack.org/project-team-guide/repository.html#retiring-a-repository

Change-Id: I5310983bed6c5b1618b7e131996e09571be9f13d