Run lower-constraints to ubuntu-bionic
The lower-constraints is supposed to test the oldest version of the
dependencies supposedly supported by a library or service. It doesn't
make sense to be running these on the newest version of Python supported
by OpenStack. Start running lower-constraints on ubuntu-bionic nodes
rather than ubuntu-focal. When we bump our minimum Python version to
e.g. 3.8, we can change this to ubuntu-focal (assuming anyone is still
working with lower-constraints by then).

Change-Id: Icaa4d8e03915d90b542b989ecb62a513125f2642
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>