wheel-cache: update openstack package for 8-stream
This old openstack package brings in centos-release-ceph-nautilus
which actually has an incorrect mirrorlist entry in it's .repo file
that tries to point it at the centos-8 repos [1].  When upstream
retired CentOS 8 this stopped responding.  Update to Xena (we don't
rely on any of these bits, it's just one or two packages we need
bindep to install for the wheel builds).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2051792

Change-Id: Ice106cfee23c514a5e5716d4368af9f62360e2f7