openafs: generate deb packages for openstack-ci PPA
This imports the latest debian/* from the upstream openafs packages,
and builds source packages meant for upload to the PPA we use for
openafs.

Over the years we have found that maintaining our own PPA for this is
useful.  We have hit bugs that require immediate patches to be applied
and rolled out to production, and having a private build facilitates
this.  Generally we hope to have no delta to the upstream packaging,
which is currently the case.

Currently we have a fairly manual process to produce and upload these
debs (once there, the source is obviously available).  This allows us
to keep a little clearer view of what we've changed or are working on.

This doesn't publish anything, yet.  We'll work on that after this;
the idea is that the publishing job will consume the artifacts this
returns.

Change-Id: I07780595331da29af9f72185cbb3499210e89c19