grafana : import and screenshot project-config graphs
This is a job that imports the graphs we have defined in
project-config/grafana to a testing grafana instance, then takes some
screenshots to validate the layout and aid review.

Operation is fairly straight-forward; a grafana container is setup
(see related work in Ibbc2b116d0c496655a7ce6bb6971e8270ac32647) and
grafyaml is used to import the dashboards from project-config.  We
query the grafana instance to find the dashboards, then use selenium
and a small helper to take screenshots of each.

It pulls graphyaml from the buildset registry as a soft dependency, so
we can also add it as an extra test to the grafyaml repo.

Change-Id: Ice0863d5a180738119d572df1871093a292126be