Address breakage due to ansible-lint 6.5
Pin ansible-lint<6.5 for now due to the
https://github.com/ansible/ansible-lint/issues/2320 regression which
ignores our skip tags.

Also add the zuul.d directory to the list of directories to avoid
linting, since those YAML files aren't Ansible files and it doesn't
seem to know not to try parsing them anyway.

While we're here, "fix" a task name not starting with an upper-case
letter, which for some reason ansible-lint now seems to think is
important. If we decide to continue using ansible-lint, it's
probably going to persist in complaining about that.

Change-Id: I46826210dfa4442af5c36d1439d1a1047ca0ba90