Make python version testing template unversioned
In Zed cycle PTG, TC agreed[1] to make python version
testing template to be unversioned. This way we can avoid
bot patches to each repo zuul yaml file to update the
template name which was missed on many repo which end up
using the old version template for many cycle. With this
unversioned template we can handle it at central place in
openstack-zuul-jobs repo.

For stable branch, we will use the branch variant to keep
same name template to continue testing the python version
applicable for that branch.

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L355

Depends-On: https://review.opendev.org/c/openstack/governance/+/854375
Change-Id: If402f9ae0ca06fec058fbc79911dd862977034ee