Add some output to the `find' command
This, alongside the "df -h" output, may provide some useful information
about the /opt content (not as precise as a `du -shxc /opt', but at
least it's already something). It can then be compared with the
existing `df -h' information located in the zuul-info/ directory. 

The goal here is to understand the actual impact of this task.

Change-Id: I201eba6ec6db9f0caaaabb0a3d15b0894125c7d6