Pin ansible-compat<4 until we upgrade ansible-lint
There was a regression introduced in ansible-compat v4 which stopped
ansible-lint from working. This was fixed in ansible-lint 6.16.0 but
we've pinned it for so long now that there are a lot of rule
violations to fix or mark for skipping. In the interim, pin back all
new versions of ansible-compat until someone has time to fix things
or rip out ansible-lint entirely.

Change-Id: I62c3ca8731f237ba19868db5d0388d130ea0f214