Prepare the job template for 2024.1 testing runtime
2024.1 cycle is going to start and we are defining
testing runtime (depends-on). This commit modifies
python job templatefor the same.

Note: this makes py3.9 and py3.10 job to run on master
periodicly (not on every change) with assumption that
running py3.8 and py3.11 jobs should be enough coverage
to run on every change and able to catch anything break
on py3.9 and py3.10 too.

Depends-On: https://review.opendev.org/c/openstack/governance/+/891225
Change-Id: Ie2b56bac8465ab6847168a09ac60c9d96d7c23e7