Continue same set of jobs running for stable/2023.2
We are near to the release of 2023.2 and release team
is starting to create the stable/2023.2 (for lib) soon.
To continue testing the same set of job (py3.9 and py3.11
n-v tox jobs, this change adds the stable/2023.2 in branch
variant for applicable jobs for 2023.2 release.

This way any early cut stable/2023.2 as well as for projects
still have master for stable/2023.2 will continue testing the
same set of jobs we do during 2023.2 release (current master).

Change-Id: I53134ef9b57e9434b8916e4fd146010074d88d59