generatepot-rst.sh: Drop UUID filtering
tools/generatepot-rst.sh has an awk filter to drop UUID from POT files,
but it is unnecessary now. It was introduced to strip UUID information
which was added by default in Sphinx <1.3 [1] in openstack-manuals.
Sphinx >=1.3 does not output UUID information by default [2],
 so is is no longer needed.

In addition, (g)awk 5.0 or later (adopted in Ubuntu 20.04 or later)
complains the current awk regexp. I think it is the time to drop it
rather than fixing the regexp.

[1] https://opendev.org/openstack/openstack-manuals/commit/993647f316fbe33a513baffb5cc4ad156355da8a
[2] https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-gettext_uuid

Related-Bug: #2035226
Change-Id: I57ce0c7d662a4cfd84d9647645a0c6ed9c56ac91