Extend the tox<4 pinning for unmaintained branches also
Tox was pinned < version 4 for all stable branches <=zed.
With new branch namespace, those stable branches moving to
unmaintained/* name[1] and endup using the latest tox version
and fail[2]. To fix that, we need to extend the tox pin for
those unmaintianed branches also.

Currently only unmaintained/yoga is proposed but soon or in
future all branches from victoria to zed are moving to
unmaintained/ so pinning tox for those branches also.

[1] https://review.opendev.org/q/topic:%22create-yoga%22
[2] https://zuul.opendev.org/t/openstack/build/201ecdc067e646868c4f17374093f328

Change-Id: I2a977fa9c57e540ded4cc5017283302106f5c807