Drop Debian Buster python wheel mirror jobs
We are in the process of dropping this node type from nodepool and our
mirrors entirely. Stop running jobs to build the python wheel
mirror/cache for this release.

Note we may need to force merge some of these depends on in order to get
this change in. Maybe we'll force merge this change instead or as well.

Depends-On: https://review.opendev.org/c/openstack/requirements/+/910247
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910241
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910242
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910243
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910244
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910245
Depends-On: https://review.opendev.org/c/openstack/requirements/+/910246
Change-Id: I0d328220b1679c8ac514bf534dbb6a908db0a2dd