Cleanup legacy CentOS 7 jobs and nodesets
We are starting work to remove CentOS 7 from Zuul and Nodepool. The
first step there is to delete everything that appears to be unused. We
do that here cleaning up nodesets and jobs that should be unused at this
point in time.

Note we keep the wheel cache/mirror build and publication jobs around
for just a bit longer. We will clean those up when we are closer to
being able to remove everything CentOS 7 related.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/910978
Change-Id: I984c494afc5ba58b50350bc9eae58260cc66f2ee