Remove CentOS 7 jobs
We're removing CentOS 7 test nodes from OpenDev and before we can do
that we need to clean up the supporting infrastructure for things like
wheel caching/mirroring which in turn depends on openafs packaging.
Enough time has passed since we announced the CentOS 7 nodes would be
removed that we should be good to start dropping this supporting
infrastructure. Go ahead and make that happen.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/912283
Change-Id: I5c56e0f081b4a4720ca59c8d2b59621fcaaad39d