Remove old infra team puppet testing
The OpenDev team is planning on removing Ubuntu Xenial in the nearish
future. One of the things still running on Xenial is our old puppet
testing. We've decided taht we'll just have to manually manage those
nodes going forward without CI as there are few of them at this point
and we really want to clean up old Ubuntus in nodepool and our mirrors.

There is a risk that this will allow things to break without detection
or otherwise create new problems, but this stuff has been neglected so
that risk was there anyway.

Depends-On: https://review.opendev.org/c/openstack/project-config/+/912309
Change-Id: I722bd1ab4fb7c0dfc2864ee55a64c4cf4dfec195