More CentOS 7 Cleanup
This removes publication of centos 7 build artifacts (like those for
openafs packages) from our centos artifacts publication job. We split
this into a separate change because centos 7 got a new kernel recently
and our existing dib images hadn't refreshed to pull that in yet which
is required for successful package updates on centos 8 and 9 stream.

Now the kernel should be updated and everything should work.

Change-Id: I369ceb1dda81f698d8b591b5f2991f99a7a823af