Drop CentOS 8 Stream fips and openafs package jobs
CentOS 8 Stream is EOL and our mirror automatically replicated the empty
package repo state for CentOS 8 Stream. The end result is that jobs
running on CentOS 8 Stream are no longer functional. In preparation for
the removal of this test node type we remove these jobs that rely on
and/or support CentOS 8 Stream in our environment.

Change-Id: I453c74ec5ffcfd8dadd77da10dc4a817342cc1bb