openstack-charms: align to Ubuntu Pythons
We've managed to end up with a mix of Python version targets for
each branch of the repositories under the OpenStack Charms project.

Python testing runtimes should be aligned to the Python base version
shipped in the Ubuntu releases each charm branch needs to support.

Re-align the main template job used for stable/2023.1 onwards with
this; this reduces the amount of test resources the projects reviews
with consume.

For the master branch (which will target Ubuntu 24.04 LTS) make the
test for Python 3.12 voting.

Change-Id: Idba042028c8100246d78c31746dd1aff98ed88f3