Remove the get_md5 parameter from ansible stat tasks
Ansible 9 removes the get_md5 parameter and it is an error to use it.
Currently openstack's zuul tenant defaults to ansible 8 but at some
point (hopefully in the not too distant future) we will set ansible 9 to
be default and we'll need to stop setting this parameter. Go ahead and
do so now.

Note this may cause ansible 8 tasks to run a little bit longer but the
impact should be minimal and makes use more future proof.

Change-Id: I9ebae82e5ab485d2d9adb4e0ef26868cbec0f80f