openstack-charms: Add charmbuild to openstack-python3-charm-jobs
The charmbuild definition is available at charm-zuul-jobs[0], this task
takes care of building and saving the resulting artifact.

Adding this job to the check and gate pipelines for the master branch
prevents from merging patches that break the build. This is restricted
to master while we monitor this is helping with the quality, and if this
proves to be sucessful, it may be considered to roll this out to the
stable branches.

[0] https://opendev.org/openstack/charm-zuul-jobs/src/branch/master/roles/charmbuild/tasks/main.yaml

Change-Id: I2086ce7cc6f841968eab21215b089792a27dfdc6