Update CentOS RPM OpenAFS build from 1.8.10 to 1.8.12.1
Recently this RPM build has been failing due to what appears to be a
mismatch in declarations for the `void abort(void);` function. Linux on
CentOS 9 Stream declares it that way but OpenAFS declars it as
`static_inline void abort(void);` and compilation fails.

Hope is that using a more up to date release of OpenAFS will allow it to
compile with CentOS 9 Streams latest kernels.

Note it is possible we may need to update the consumption side of this
packaging; however, I'm not finding any obvious cases where we hardcode
the rpm package version [0]. I hope that means it is largely automatic.

[0] https://codesearch.opendev.org/?q=1%5C.8%5C.10&i=nope&literal=nope&files=&excludeFiles=&repos=

Change-Id: I6f4cda181a1d18895ad39e00ccf1c2639cfe2154