Make scripts DCO compliant
The OpenStack project will be switching from the OpenInfra
Individual Contributor License Agreement to the Developer
Certificate of Origin, which requires that commit messages include a
Signed-Off-By trailer. Any commits generated by automated tooling
will need to comply with this requirement as well. Simply including
git commit's -s option on the command line should be sufficient to
meet this requirement.

Change-Id: I8f94080f2d5ee64be3bbb14b676f9029b9693b00