Remove bionic-arm64 wheel build jobs
This is cleanup necessary as there are no bionic arm64 images built by
nodepool in zuul. As part of the transition off of nodepool we
need to stop relying on these node types.

We did do some checking and it doesn't appear that arm64 jobs really got
serious (doing useful work) until python38 focal jobs. This is believed
to not have a major impact as a result. Either way the release is long
out of support and is due to be cleaned up.

Now that we have turned off nodepool, undefined labels have become
configuration errors.  Our earlier "dummy" xenial nodesets are
now errors.  The easiest way to correct them (aside from removing
them) is to set them to be zero-node nodesets.

Depends-On: https://review.opendev.org/c/openstack/requirements/+/954767
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954770
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954771
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954772
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954773
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954774
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954775
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954776
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954777
Depends-On: https://review.opendev.org/c/openstack/requirements/+/954779
Change-Id: I45983382bcede34312350cf2b127072883441c3b
Signed-off-by: James E. Blair <jim@acmegating.com>
Signed-off-by: Clark Boylan <clark.boylan@gmail.com>