Make dummy ubuntu-xenial nodeset
This is necessary because we've removed ubuntu-xenial as part of the
nodepool in zuul migration but there are many jobs that still want to
use an ubuntu-xenial nodeset. Give them a nodeset that will result in
NODE_FAILURE which should prompt people to clean up these broken jobs.

Note they were already broken this just makes them broken while having
valid zuul configuration.

Change-Id: I28da9c6c7ca5599a29a610b57d3c81ee45dcdc4c
Signed-off-by: Clark Boylan <clark.boylan@gmail.com>