Remove bionic nodeset from unit and functional test jobs
ubuntu-bionic nodeset will be removed from zuul's nodepool, hence it
cannot be used for tox-py27, tox-py36 and tox-py37 jobs and the same
functional jobs. This patch removes the use of ubuntu-bionic nodeset
where possible and adds the use of pyenv instead.
However, tox-py36 job needs to keep ubuntu-bionic and force the usage
of Ansible version 9 (Zuul will drop the support of Ansible 9, and it
can only be run by explicitly specifying it to be used, otherwise the
job just fails without any logs).

Change-Id: I519b5b910acc4bd1c197acb46d33142c5be946ef
Signed-off-by: Elod Illes <elod.illes@est.tech>