Convert py27 job to ubuntu jammy
Ubuntu Jammy is the last Ubuntu LTS with python2.7 packaged. However, it
uses different package names than Bionic so we need to update to match.
Make this conversion because Bionic is on its way out and if we're going
to keep running 2.7 tests we need to do so on a newer platform.

Change-Id: I1529bfea454a4d16c1a3d1b6db730a0accdf3d40
Signed-off-by: Clark Boylan <clark.boylan@gmail.com>