Drop Bionic based jobs from project templates
Python3.6 and 3.7 jobs run on Bionic. Unlike python2.7 there is not good
alternative for running 3.6 and 3.7 jobs. With Bionic on its way out
(zuul will drop support for it via Ansible updates) we should start
cleaning up our reliance on this platform. Since we can't run 3.6 and
3.7 jobs elsewhere we simply remove them.

The first step in that removal is clearing the jobs out of project
templates. Do that here.

Change-Id: Ifcfba03d1fa73e31e9f60c332c40365b6590b53e
Signed-off-by: Clark Boylan <clark.boylan@gmail.com>