Switch to building releasenotes with tox
The current implementation of this job manually installs and calls
'reno' and it relies on 'doc/requirements.txt' files to install other
dependencies. This is a problem as we explore moving dependency
information into 'pyproject.toml'.

Rather than doing this, switch to using tox for release notes as we
do for docs, api-ref, and api-guide. This should have zero impact given
the 'releasenotes' tox testenv is found in virtually all projects.

Change-Id: I98127c5976332770fa6b42513714a05df6587421
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>